+++ lua/babel/init.lua
@@ -1,7 +1,9 @@
 ---@class Babel Main plugin module
 ---@field setup fun(opts?: BabelOptions) Initialize plugin
 ---@field translate fun(text: string) Translate text
+---@field translate_range fun(line1: number, line2: number) Translate line range
 ---@field translate_word fun() Translate word under cursor
+---@field repeat_last_translation fun() Repeat last translation
 
 ---@type BabelConfig
 local config = require("babel.config")
@@ -38,6 +40,14 @@ function M.translate(text)
   translate.translate(text)
 end
 
+---Translate line range
+---@param line1 number Start line (1-based)
+---@param line2 number End line (1-based)
+function M.translate_range(line1, line2)
+  local text = utils.get_line_range(line1, line2)
+  translate.translate(text)
+end
+
 ---Translate visual selection
 function M.translate_selection()
   local text = utils.get_visual_selection()
@@ -48,6 +58,11 @@ end
 function M.translate_word()
   local text = utils.get_word_under_cursor()
   translate.translate(text)
+end
+
+---Repeat last translation
+function M.repeat_last_translation()
+  translate.repeat_last()
 end
 
 return M